+++ src/components/graph/selectionToolbox/DeleteButton.vue
@@ -21,6 +21,7 @@ import { useI18n } from 'vue-i18n'
 
 import { useSelectionState } from '@/composables/graph/useSelectionState'
 import { Positionable } from '@/lib/litegraph/src/interfaces'
+import { useCanvasStore } from '@/renderer/core/canvas/canvasStore'
 import { useCommandStore } from '@/stores/commandStore'
 
 const { t } = useI18n()
