+++ tests-ui/tests/performance/transformPerformance.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it } from 'vitest'
 
-import { useTransformState } from '@/renderer/core/layout/useTransformState'
+import { useTransformState } from '@/renderer/core/layout/transform/useTransformState'
 
 // Mock canvas context for testing
 const createMockCanvasContext = () => ({
